+++ bin/ota.js
@@ -11,6 +11,7 @@ program
   .description(description)
   .version(version)
   .command('track', 'Track the current terms of services according to provided declarations')
+  .command('apply-technical-upgrades', 'Apply technical upgrades: regenerate versions from existing snapshots with updated declarations/engine')
   .command('validate', 'Run a series of tests to check the validity of terms declarations')
   .command('lint', 'Check format and stylistic errors in declarations and auto fix them')
   .command('dataset', 'Export the versions dataset into a ZIP file and optionally publish it to GitHub releases')